Added filter and moved pipes/directives to App folder
