+++ Angular/angular-mentoring/src/app/app.module.ts
@@ -1,44 +1,46 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
-import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { HeaderComponent } from './header/header.component';
-import { FooterComponent } from './footer/footer.component';
+import { AppRoutingModule } from './app-routing.module';
 import { BreadcrumbsComponent } from './breadcrumbs/breadcrumbs.component';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { CourseItemComponent } from './course-item/course-item.component';
 import { CourseListComponent } from './course-list/course-list.component';
+import { FooterComponent } from './footer/footer.component';
+import { HeaderComponent } from './header/header.component';
 
-import { MatCardModule } from '@angular/material/card';
 import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
 import { MatInputModule } from '@angular/material/input';
 
+import { BorderColorDirective } from './directives/border-color.directive';
+import { DurationPipe } from './pipes/duration.pipe';
+import { FilterPipe } from './pipes/filter.pipe';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
-import { BorderColorDirective } from '../directives/border-color.directive';
-import { DurationPipe } from '../pipes/duration.pipe';
-import { OrderByPipe } from '../pipes/order-by.pipe';
+import { OrderByPipe } from './pipes/order-by.pipe';
 
 @NgModule({
   declarations: [
     AppComponent,
-    HeaderComponent,
-    FooterComponent,
+    BorderColorDirective,
     BreadcrumbsComponent,
     CourseItemComponent,
     CourseListComponent,
-    BorderColorDirective,
     DurationPipe,
+    FilterPipe,
+    FooterComponent,
+    HeaderComponent,
     OrderByPipe,
   ],
   imports: [
     AppRoutingModule,
     BrowserAnimationsModule,
     BrowserModule,
+    FontAwesomeModule,
     MatCardModule,
     MatInputModule,
     MatButtonModule,
-    FontAwesomeModule,
   ],
   providers: [],
   bootstrap: [AppComponent]
