+++ Angular/angular-mentoring/src/app/pipes/duration.pipe.ts
