+++ Angular/angular-mentoring/src/app/pipes/filter.pipe.ts
@@ -1,0 +1,20 @@
+import { Pipe, PipeTransform } from '@angular/core';
+
+@Pipe({
+  name: 'filter'
+})
+export class FilterPipe implements PipeTransform {
+
+  transform(values: any[], filter: string): any {
+    if (filter != undefined) {
+      filter = filter.toLowerCase();
+      values = values.filter(value => {
+        const { title, content } = value;
+        return title.toLowerCase().includes(filter) || content.toLowerCase().includes(filter);
+      });
+    }
+
+    return values;
+  }
+
+}
